trying an architecture to ease data flow in extension
Reworked initial architecture:

- background is now a complete es6-enabled env.
- content is now a complete es6-enabled env.
- some common code between the background, content and panel envs
is put in a new "common" dir.
- helpers/themes jsons are put together in a "data" dir

Each env (background, content, panel) has an internal redux store.
To communicate, a "sync" system enables stores to talk to each other by
defining action types to receive or send to others.

Mostly, the "container" related actions are shared. This is because:
- the buttons to change iframe state are part of the "panel"
- the code having access to the iframe dom element is in "content"
- the code being able to generate a popup window is the "background"

So, each part tells when it wants to do stuff, and we react to those
demands by watching for actions through sagas.