Synchronizing stores using middlewares
Each part of the code still has its own store, but they are now
synchronized using custom middlewares, that broadcasts every action they
handle to the other stores.

This allows for synchronisation of actions across stores, but does not
guarantee each store's state will be the same. For this to happen, one
would have to mount the same reducers and sagas on each store, so that
actions are handled exactly the same way everywhere.