Asynchronous loading of mappings
Mapping are not build into JS files anymore, but requested on the fly.
This removes the hard coded imports to load initial state, and
strengthens the code, making it work even when no data is loaded.
This is also a first step to be able to retrieve mappings from external
APIs.