fixing panel being empty or not depending on how we open the panel
linked to the fact we have one global store for all tabs ; if you opened
the panel through the browser extension button, then reloaded the page,
then clicked again the browser extension button, the background store
still thought the panel was open, while it was not. So it never loaded
data again. I think.