fixing options not showing in Firefox
option page was broken in FF because option tab object is different in chrome and FF:
- in chrome, the option page has no "tab" object in the messages it sends
- in firefox, the option page has a "tab" object with the url "about:addons" in the messages

Previously, we checked for the presence of a tab & tab.id to guess it was the option page or not.
We know obviously know we couldn't rely on that, now we rely on the options page url directly.