injecting JS content scripts when clicking the action button
This prevents content scripts to try getting broken initial states.
Before, when clicking on the toolbar extension button, we created in the background a state for the
current tab.
But we required the current tab's state in the content scripts, and they were loaded directly with the
page.
So, if we didn't click the button, we had a wrong initial state and we had to do weird empty promise
catches.

Now the content script is always loaded *after* a tab's state has been initiated.

Current drawback is perfs seem to be impacted by this, we'll have to check that.