fix a bug when closing/opening toolbar multiples times on same tab - #84
The visible problem was that the helpers were applied twice after reopening the toolbar.
It seems the problem was that we loaded content scripts each time we opened
the panel, whereas we needed to do what we do now, which is loading the scripts
only the first time we click. Now every action is only triggered once as intended.