Fixed a bug with message passing when the extension runs in a popup
Dereferencing an instance on its tab id was causing problems in case
helpers were communicating with widgets.
Helpers sent messages identified with the tab id they were running in,
but the background pool couldn't find the instance associated with this
tab.
Now, we're just keeping the reference so these messages can arrive to
their intended instance.