panel minimizing now actually works!
Code is a bit weirdin ResizeHandle, this is done in order to keep the
actual DOM elements and just hide/show them on demand, instead of
rerendering everything on minimize/maximize. I'm sure we could do better
than that though.

Current behavior:

* closing (via extension btn or close icon in the extension) hides the
toolbar + removes helpers
* minimizing (via minimize icon) hides the toolbar without removing
helpers. A little toolbar icon appears in the top right corner of the
page to open back the app.